+++ .gitignore
@@ -19,3 +19,4 @@ test/output
 test/logs
 test/*.json
 test/ft
+requirements_tmp_for_setup.txt
