+++ gui.sh
@@ -17,6 +17,12 @@ cd "$SCRIPT_DIR"
 source "$SCRIPT_DIR/venv/bin/activate"
 
 # If the requirements are validated, run the kohya_gui.py script with the command-line arguments
-if python  "$SCRIPT_DIR"/tools/validate_requirements_unix.py -r "$SCRIPT_DIR"/requirements_unix.txt; then
-    python "$SCRIPT_DIR/kohya_gui.py" "$@"
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    if python  "$SCRIPT_DIR"/setup/validate_requirements.py -r "$SCRIPT_DIR"/requirements_macos.txt; then
+        python "$SCRIPT_DIR/kohya_gui.py" "$@"
+    fi
+else
+    if python  "$SCRIPT_DIR"/setup/validate_requirements.py -r "$SCRIPT_DIR"/requirements_linux.txt; then
+        python "$SCRIPT_DIR/kohya_gui.py" "$@"
+    fi
 fi
