+++ pyproject.toml
@@ -1,0 +1,24 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "kohya_ss"
+version = "1.0.3"
+description = "A GUI wrapper for kohya-ss SD scipts enabling LoRA training with an easy-to-use web application."
+authors = [
+    {name = "bmaltais", email = "bernard@ducourier.com"},
+]
+readme = "README.md"
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "License :: OSI Approved :: Apache Software License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+]
+
+[tool.poetry]
+license = "Apache-2.0"
+
+[tool.setuptools.packages.find]
+where = ["library"] # We have to explicitly tell build tools where to look
