+++ requirements_macos.txt
@@ -1,4 +1,4 @@
-accelerate==0.19.0
+accelerate==0.15.0
 albumentations==1.3.0
 altair==4.2.2
 bitsandbytes==0.35.0
@@ -8,20 +8,15 @@ easygui==0.98.3
 einops==0.6.0
 fairscale==0.4.13
 ftfy==6.1.1
-gradio==3.23.0; sys_platform == 'darwin'
-gradio==3.32.0; sys_platform != 'darwin'
-huggingface-hub==0.13.3; sys_platform == 'darwin'
-huggingface-hub==0.13.3; sys_platform != 'darwin'
+gradio==3.23.0
+huggingface-hub==0.13.0
 lion-pytorch==0.0.6
 lycoris_lora==0.1.6
 opencv-python==4.7.0.68
-prodigyopt==1.0
 pytorch-lightning==1.9.0
 rich==13.4.1
 safetensors==0.2.6
-tensorboard==2.10.1 ; sys_platform != 'darwin'
-tensorboard==2.12.1 ; sys_platform == 'darwin'
-tensorflow==2.10.1; sys_platform != 'darwin'
+tensorboard==2.12.1
 timm==0.6.12
 tk==0.1.0
 toml==0.10.2
