+++ setup/debug_info.py
